+++ database/migrations/2025_11_13_212048_create_suscripciones_table.php
@@ -11,18 +11,20 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('suscripciones', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->unsignedBigInteger('servicio_id');
-            $table->foreign('servicio_id')->references('id_servicio')->on('servicios')->onDelete('cascade');
-            $table->enum('estado', ['activa', 'cancelada', 'expirada'])->default('activa');
-            $table->timestamp('fecha_inicio');
-            $table->timestamp('fecha_fin')->nullable();
-            $table->timestamp('proximo_cobro')->nullable();
-            $table->json('metodo_pago')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('suscripciones')) {
+            Schema::create('suscripciones', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->unsignedBigInteger('servicio_id');
+                $table->foreign('servicio_id')->references('id_servicio')->on('servicios')->onDelete('cascade');
+                $table->enum('estado', ['activa', 'cancelada', 'expirada'])->default('activa');
+                $table->timestamp('fecha_inicio');
+                $table->timestamp('fecha_fin')->nullable();
+                $table->timestamp('proximo_cobro')->nullable();
+                $table->json('metodo_pago')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
